[distr, distrDoc, distrMod in branches] triggered by mail by Achim Zeileis, we replaced obsolete URLs to JSS papers by respective DOIs
(Mail by ACHIM:)
Dear Peter Ruckdeschel,

the Journal of Statistical Software (JSS, https://www.jstatsoft.org/) recently migrated to a new server and editorial system, resulting in a change of the URLs being used for publications. Hence we checked all CRAN packages using JSS URLs in the documentation or citation files etc. This includes some of your packages: distr, distrDoc, distrMod.

In general we recommend to use DOIs instead of URLs to link to JSS publications. These use the following pattern for articles: 10.18637/jss.vXXX.iYY where XXX is the three-digit volume and YY the two-digit issue. (For code snippets a "cYY" instead of "iYY" is used.) The DOIs are also shown on the web pages of the JSS articles.

For including these in a package you typically use:
- \doi{...} markup in .Rd files
- <doi:...> in DESCRIPTION/Description fields
- bibentry(..., doi = ...) in CITATION files (or citEntry)

We would recommend to change all JSS references in your package correspondingly (even if redirections for the URLs are still working). The corresponding files in the package are:
distr/inst/CITATION
distrDoc/inst/CITATION
distrMod/man/0distrMod-package.Rd
distrMod/inst/CITATION

Thanks for your consideration - and for referring to work published in JSS!

Best regards,
Achim Zeileis
(JSS co-editor-in-chief)

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/distr@1360 edb9625f-4e0d-4859-8d74-9fd3b1da38cb